to_number: Fix imprecise conversion for small floating-point numbers
[why]
The conversion of 1.17549e-38 to float was off by up to
6 ULPs.

[how]
Convert the after-dot part of the number to an unsigned
integer first, then multiply it with a float power of 10.
This algorithm is only used if the after-dot part is
actually short enough to fit into an unsigned integer,
otherwise we revert to the previous algorithm.