lint: Exclude optional.h
[why]
As with string_view the backport of optional is full of linter warnings.
As we do not want to fix these we should also not see them.

[how]
Exclude (i.e. do not include) the optional header.

I hope that the 'grep' regex works on all platforms. Usually I use egrep
or grep -e, but that might prove even more brittle.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>