Add Ryu to GUL
[why]
The question came up whether we want to have a Grisu implementation in
GUL. The more logical solution would be the {fmt} library. But some
experimentation showed that the Grisu implementation there is hard to
access directly.

If just a Grisu-like functionality is needed Ryu might be even better.
And it has a direct and easy interface.

[how]
Forked Ryu on Github and added meson and catch2 support:
https://github.com/Finii/ryu

In that repository you can also find a very basic example to play around
(example.cc) done by me.

Used that to incorporate it into GUL.

[note]
If we would decide to really pull this (which I do not expect) we need
to add copyright information stuff, add Doxygen, ....

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>