Add systematic check if headers are self-contained
[why]
SF.11: Header files should be self-contained

Usability, headers should be simple to use and work when included on their own.
Headers should encapsulate the functionality they provide. Avoid clients of a
header having to manage that header’s dependencies.

A header should include all its dependencies. Be careful about using relative
paths because C++ implementations diverge on their meaning. Enforcement

Note: Failing to follow this results in difficult to diagnose errors for
clients of a header.

[how]
A test should verify that the header file itself compiles or that a cpp file
which only includes the header file compiles.

[note]
Also try to instantiate all template variants that could be relevant.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>