to_number: Make sure that the correct overload of pow() is chosen
[why]
The standard guarantees the overload float std::pow(float, int),
but in our previous implementation, the second parameter was
a size_t. On VS2019, this chose an overload returning double,
which in turn generated a warning because a double was
assigned to a float variable.

[how]
Add a static_cast from size_t to int.