to_number: Use long double internally for floating-point conversions
[why]
On VS2019, the previous algorithm fails the test for converting
"1.17549e-38" to float with an error of 1 ULP or less. This is caused
by a huge error of std::pow(10.0f, -39).

[how]
Use long double for all internal calculations, then cast back to the
target format just before returning the value.