to_number: Expand test to try type 'long double'
[why]
The function to_number() is not limited to float and double, so some
user might convert long double. But there are no tests at all.

[how]
Because catch2's matcher does not support long double we use GUL's
matcher.

[note]
On MSVC the types double and long double are the same 64 bit (imho).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>