Correctly handle macro GUL_USING_STATIC_LIB_OR_OBJECTS
[why]
The documentation specifies that GUL_USING_STATIC_LIB_OR_OBJECTS has to
be (merely) defined to enable some option. The previous code depended on
it to be defined to something which is not null. This is a
contradiction.

[how]
After using the preprocessors #ifdef in #elif branches usually a
defined() operator is needed to get the desired effect. #ifdef add the
defined() operator only to the first #if call.

[note]
Old macro definitions that activated the feature
 #define GUL_USING_STATIC_LIB_OR_OBJECTS
 #define GUL_USING_STATIC_LIB_OR_OBJECTS 1
Old macro definitions that deactivated the feature
 #undef GUL_USING_STATIC_LIB_OR_OBJECTS
 #define GUL_USING_STATIC_LIB_OR_OBJECTS 0

The last case (defining it as being zero) is now lifted to the 'activated'
section. This new behavior is the same as with GUL_COMPILING_SHARED_LIB,
where a definition to zero is still a definition and activates the
feature.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>