meson: Hide symbols per default with GCC/clang
[why]
For compatibility with Windows we have commit
  c4842f8 Add GUL_EXPORT macro to handle DLL symbol export and import on Windows

If we run tests on a GCC ro clang based build we want to see if the
symbol exporting has been done right immediately - not only when done on
a MSVC platform.

[how]
Turn on -fvisibility=hidden on GCC and clang.
Use appropriate macro to make certain symbols visible.

[note]
This als prevents leaking internal symbols into the GCC/clang world's
dynamic libraries.

[note]
The feature used has been introduced in meson version 0.48.
It will be ignored in older meson versions.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>