internal: Indent nested preprocessor commands
[why]
The readability can be improved if preprocessor commands are indented if
they are nested.

Using 'if defined(...)' rather than 'ifdef ...' when an #elif branch
exists can prevent wrong assumptions about what is tested.

[how]
We keep the '#' in the first column to mark the line clearly as
preprocessor line.

[note]
Also added vimline to enable easy indentation.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>