VS2019: Make make_vs independent of current working dir
[why]
The previous version had to be called from the repository root directory
to get all the directories straight. This puts some burden on the user,
because
 - one has to be in a specific directory to do make_vs
 - one has to call make_vs with at least a partial path ('tools/')

[how]
Determine the absolute path of the repository when the make_vs is called
and use absolute paths in all subsequent (e.g. meson, ninja, ...) calls.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>