VS2019: Automatically find Visual Studio installation
[why]
People need to change the make_vs script to put the path of their
Visual Studio installation in. This is an inconveniance.

[how]
Try to find the most recent Visual Studio, one of
  VS2019, VS2017, VS2015; Enterprise, Professional, Community
at the default installation path.

Users can specify a differnt location if they want, this prevents
the automatism.

Usually meson is installed in the %path%, so the setting for its
absolute path is also optional.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>