to_number: Use correct vimline
[why]
The previous vimline sets noexpand tabs, but we do not want to have
tabs; vim should expand them to spaces.

[how]
Use the vimline used in (almost) all other header files.

This drops both the textwidth and the indent type, that were set
explicitly in the former file.

[note]
This is quite a big and ridiculous commit, for just correcting the
vimline. I see no real reason to have this in a separate commit, as
noone ever will want to know why and how the vimline has been changed.
But of course from a pedantic standpoint it is absolutely essential to
separate this commit from commits that change the code.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>