tests: Use type_t type instead of auto and literal
[Why]
Under vc2017 the compilation of the sliding buffer test fails with the
following error:

> [...]
> FAILED: tests/59830eb@@libgul-test@exe/test_SlidingBuffer.cc.obj
> cl @tests/59830eb@@libgul-test@exe/test_SlidingBuffer.cc.obj.rsp
> tests\test_slidingbuffer.cc(1129): error C2398: Element '1': conversion from 'const int' to 'unsigned int' requires a narrowing conversion
> [...]

[How]
There are multiple ways of solving this issue, e.g. one could define
and use a size_t suffix literal as proposed [1] for C++23, since it
is C++ stylish :-) But since it does not really improve the
readability, we simply drop the auto instead and specify the instead.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>
--
v2: Drop size_t suffix literal

[1] http://open-std.org/JTC1/SC22/WG21/docs/papers/2019/p0330r5.html