to_number: Use floating exponent with std::pow()
[why]
This commit is a variation of the reverted commit
  c7af61b to_number: Use long double internally for floating-point conversions

Here is the commit message of that commit, which is also the aim of this
commit:

  On VS2019, the previous algorithm fails the test for converting
  "1.17549e-38" to float with an error of 1 ULP or less. This is caused
  by a huge error of std::pow(10.0f, -39).

  [how]
  Use long double for all internal calculations, then cast back to the
  target format just before returning the value.

On MSVC the integer exponent overload (i.e. the general arithmetic type
overload) does not create subnormal results. But the floating exponent
overload works.

[how]
Use the correct std::pow() overload.

The way we construct the result does not handle values near of lower
than numeric_limits<>::min (i.e. subnormal values) gracefully. The error
increases and so tests fail. So these tests are disabled (for now).

[note]
Not all tests succeed with this commit.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>