to_number: Make float overflow tests systematic
[why]
There are some tests for conversion to float that check minimal and
maximal representable values together with subnormal values.

The tests assume a certain accuracy of the type float.
The tests do not check the same for double or long double.

[how]
Generate the string to be converted from the type traits of the
destination type.
Check this for all three (float, double, long double) types.
Check subnormal value representation more thoroughly.

[note]
This test breaks the current implementation.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>