SlidingBuffer: Correct noexcept specification
[why]
The move assignment and move construction of a SlidingBuffer object can throw,
if the contained elements (of ElementT) can throw on move.

[how]
We need to check the noexcept specification of ElementT to deduce the
noexcept specification of SlidingBuffer.

Or to be more exact we need to check the underlying container's noexcept
specification (which in turn checks ElemenT's).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>