to_number: Add test that converts arbitrary numbers
[why]
To really see how well the algorithm fairs we need to test against
arbitrary numbers. Expected is a round trip (number to string to number)
conversion with only a limited error in all cases.

[note]
THIS FAILS. MISERABLY. On all implementation attempts, but the standard
library based one (commit 0b4b68ac51). Meaning on the original
implementation by Lars and all intermediate ones including the last done
by me.

We can try to remedy this; on the other hand maybe we want to drop the
design goals and use that commit?

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>