num_util: Correct within_ulp() for subnormal values
[why]
The previous implementation bailed out if the input numbers were
subnormal values - and just returned 'is equal enough' for any ULP
distance.

[how]
If the input number with the bigger absolute magnitude is a subnormal
value (i.e. is represented with reduced resolution) this means
 - the smaller magnitude value is also subnormal, or zero
 - the code precious code breaks because the resolution is too low

Because of the second point the previous implementation did not work on
subnormal numbers but returned TRUE in any case.

Now we add code that handles the subnormal cases correctly.

It also handles the other special floating point numbers (i.e. NAN and
INF) correct.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>