to_number: Yet another to_number() implementation
[why]
Shipped gul::to_number() still broken for a lot of tests.

[how]
Try to implement this as far according to Lars' design goals as
reasonably possible. So:

float and double: constexpr, noexcept, not allocating
long double: not constexpr, not noexcept, allocating

This asserts that the mantissa of double fits into uint64_t.

Furthermore reduces the number of pow() calls to 1. Assembling the final
number in multiple steps adds too much rounding issues anyhow.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>