to_number: Change for MSVC: Use type to specialize
[why]
Microsoft Visual C++ 19 seems to be unable to resolve more complex partial
template specialisations and warns that they are the same and can not be
redefined.

[how]
Tried a lot of different approaches to make MSVC recognize the templates,
to no avail.

Instead of working with the concrete numeric resolutions that need to be
compared to select the right algorithm we use the 'typical' implementation
size expectations instead. In this way we can simply specialize the template
on type rather than combined type traits.

Now float and double target types are always converted with the noexcept
template (our own code) and only long double is converted with strtod() et
al.
Before we checked if the requested target type's mantissa could be represented
with the FloatConversionIntType; which could vary depending on compiler and
platform.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>