to_number: Handle 'long double' constexpr if possible
[why]
The usual implementation of 'long double' on x86/x64 CPUs uses a 80 bit
representation. Operations for that floating point numbers are available
in hardware on that platform, so that is often chosen by compiler
implementers.

If that format is used the mantissa has 64 bits. This exactly fits into
out integral mantissa buffer of uint64_t. So it is possible to convert
numbers with our current implementation, with maybe an additional
rounding error (because only numeric_limits<long double>::digit10
characters of the input are considered and not
numeric_limits<long double>::max_digit10, which would be strictly needed.

Because we anyhow have a rounding error in our algorithm (because
subsequent digits are not considered and rounding is always TO_NEAREST
instead of a more fitting value), this makes the conversion not worse.

[how]
Use 'our' function to convert the number for all types whose mantissa
fits into uint64_t (this includes 80 bit long doubles). For all other
cases still use std::strtold().

[note]
On Darwin we always fall back to std::strtold() because std::pow() with
long double numeric type calculates strangely here (maybe it uses just
double internally?).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>