to_number: Improve performance by removing pow10()
[why]
The previous implementation with a table of pow10() values needs to
either inline the table (needs space), copy the table (needs time), or
reference a global table in the library (needs cachelines).

[how]
We can calculate the pow10() values on the fly. This can be cheaper than
getting the value from somewhere else. Especially when we iterate over
the characters already.

The magnitude of the converted digits is calculated while the digit
sequence is translated into an (integer) value. That branch is removed
by the optimizer if it is not requested by the template parameter.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>