to_number: Remove warning on MSVC
[why]
MSVC warns about 'possible loss of data' when converting from
long double to float. Although this is true, it will never
happen, because it is disabled in a if-statement that guards
the instantiation of this function.

Properly it should have been a 'if constexpr' and this warning
would not be there at all.

This all is kind of a tradeoff which parts of TMP the various
compilers speak :-/

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>