string_view: Make string_view constexpr-able
[why]
Although string_view should be constexpr it is in fact not. We had
several workarounds for this, instead of fixing the cause.

[how]
The defect lies in std::char_trait, that is not constexpr ready with our
c++14 standard libraries.

We inject constexpr functions into the traits used by the string_views.

Further we need to adapt the conversion from/to std::string, because
usually the traits are part of the type and are handed over unchanged.
But std::string does not work with our traits if the convenience aliases
are used (which is almost always the case).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>