type_name: Fix usage of substr() and find_last_of()
[why]
With GCC 5.4 we have difficulties working with substr() in the
constexpr context. That compiler dislikes the throw (which is
required by definition), although throw is allowed in constexpr
functions by the standard (it just must not throw).

With cl 19.22 we have difficulties working with find_last_of()
in the constexpr context. For some reason the reverse iterators
can not be used.

[how]
Reformulate the string manipulation in a way that it avoids both
problematic functions.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>