Explicitly state that at least Visual C++ 2017 is needed
[why]
Building out of the box with Visual C++ 2015 (or prior) is impossible
because we use C++14 language features, that are not implemented
in that versions of Visual C++. They were first implemented with
Visual C++ 2017 15.7.

This is namely
- Two phase name lookup
- Extended constexpr (not only one return statement) (N3652)
- Aggregate member initializers (N6353)

and maybe more.

The needed changes in the project (I tried some) are rather voluminous
and result sometimes in ugly workarounds.

[how]
Also remove the (anyhow defunct) support for 2015 from the
makefile_vs.bat.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>