to_number: Improve readability of NaN and Inf parsing func
[why]
A nested optional<> is not very easy to grasp.

[how]
Rename function from check_inf_nan() to parse_inf_nan.

We have 4 cases to handle
- We parsed the string meaning to be NAN
- We parsed the string meaning to be INF
- We found a syntax error in the string
- We found nothing special (i.e. string can be a valid number)

The first 3 cases can be used directly from the (sole) function that
call parse_inf_nan() as return values. It returns an optional, that is
not filled on syntax/parse error and otherwise holds the value.

So we return an optional that can be reused by the calling function in
the first three cases, and an extra flag that indicates we found nothing
(special).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>