to_number: Drop all uses of optional<>{}.value()
[why]
At the points we use an optional:
We always check for has_value() or operator bool before we access the
value of the optional. So there is no need to check it again by
explicit use of value().

And besides... If the check done with value() fails we would throw,
something we promised not to do by our noexcept specification ;->

[how]
Use operator* instead, that does not check has_value.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>