tests: Fix time critical tests on Windows
[why]
On Windows the tests that involve timing (i.e. gul::sleep etc), very
often fail with a higher than accepted delay.

The specification of sleep (or the underlying library functions) is to
wait AT LEAST the specified amount of time. It can take longer,
especially in a system with many running threads due to scheduler
decisions.

Obviously running the test through make_vs.bat keeps the testing thread
in a very low (idle) scheduler priority and we are very succeptible
to other cioncurrent activity on the machine.

[how]
Increase the thread priority to slightly above normal.

[note]
Also fixed allowed time deviation in one place, where we sleep two
consecutive times, so we need double the lenience. That went unnoticed
on less time critical systems, but became a problem on Windows.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>