substring_checks: Fix starts_with() on GCC 5.4
[why]
With GCC 5.4 we have difficulties working with substr() in the
constexpr context. That compiler dislikes the throw (which is
required by definition), although throw is allowed in constexpr
functions by the standard (it just must not throw).

Somehow this compiler has been missing in my manual CI checks when
we did commit
  cc3ce7ac Revert "substring_checks: Replace own implementation with string_view members"

[how]
Reformulate starts_with() in a way that it does not need substr().
Which means not using compare(x, y, z) but only compare(z).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>