Add bit manipulation functions
[why]
We often need to do bit manipulation when working with hardware. It
gives flags and stuff as bits in values that are transferred.

To access the bits usually one uses enums or equal language mechanisms
to abstract the actual bit values away. But even setting the enums up
often involves 'bit counting'.

A often used solution are macros like this:
    #define BIT(x) (1 << (x))
    #define IS_BIT(v, b) (((v) & BIT(b)) == BIT(b))

Some reasons for a real C++ solution to this problem are
- Macros undermine C++ language features (Core Guidelines ES.31)
- Parameter v undergoes integral promotion which might be unexpected
- Bits outside sizeof(int) can't be accessed but no error is raised
- Even the MSB of int can't be accessed, no error raised

[how]
Substitute function templates bit_set() and bit_test() are implemented
in a type save manner.

For completeness, but usually unneeded (because we want to work with bit
pattern enums or equivalents) bit_clear() and bit_toggle() are provided
as well.

[note]
Example for reason 3:

    // Assuming sizeof(long int) * 8 > 60
    // Assuming sizeof(long int) > sizeof(int)
    int bit = 60;
    long int v = 1;
    v << bit;
    if (IS_BIT(v, bit)) ....

If 'bit' is not given directly in the macro no compiler warns, even in
constexpr situations.
The result of the test is undefined (and usually wrong).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>