bugfix: Do standalone test on string_util.h
[why]
The header string_util.h is a normal GUL header, not a backport and
nothing else special about it. Users should be able to include it
without need for any additional other headers (it shall bring all its
own dependencies).

[how]
Just the test is not done. The test does not fail. It is unclear why the
header file has been sorted into the 'do not test' section. Possibly a
mistake.
Move the file to the appropriate section.

[note]
The backports string_view.h and optional.h are in fact also able to
standalone and should be. I'm not sure why we (I?) decided to exclude
them from the tests.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>