bit_manip: Remove undefined behavior
[why]
It is known and documented that the bit_*() function do not work if the
requested bit is outside the range of bits that the targeted type has.
E.g. if you want to set bit 18 in int16_t.

If for some reason working code, that possibly does not know the actual
type because it is templated code, or because it is compiled on a 32 bit
platform rather than the developer's envisioned 64 bit platform,
accessed bit that are not there this will be noticed only late if ever.

[how]
The function now asserts that the requested bit is in the valid range.
With this the program will fail early and with a useful debug hint.

If the added cost of the branch is too much the libgul user could simply
define NDEBUG (or use meson's -D b_ndebug=release).

[note]
I had the assert in the first version of these functions, but removed
them laster with the thought: "it's documented, so it's the users
fault". But today I watched
  CppCon 2019: Satabdi Das “A Series of Unfortunate Bugs”
where exactly this happened in production code after 10 years of code
usage... This can be prevented for so little cost.

And the intended (hmmm) use case for the function set is not in hot code
but rather in constexpr compile time enum setup.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>