to_number: tests: Add superfluous zeros test and reduce test runtime
[why]
The algorithm guarantees the correct conversion of numbers
like 000000000000000000000000001.000000000000000000000,
but this was not ascertained by the previous tests. Also, the
"random round time conversion" test took several seconds
on various systems.

[how]
Add tests for the conversion of some string literals with lots
of leading and trailing zeros. Reduce the number of random
tests from 100 000 to 5 000.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>