statistics: Remove compiler warning
[why]
With a newer compiler (here: GCC 9.2.1) we get:
warning: narrowing conversion of ‘...’ from ‘int’ to ‘double’ [-Wnarrowing]

[how]
Cast instead of implicit conversion.

[note]
Further down we already use the static cast. I believe the original code
intention was to return ElementT as ResultT. This explains the code and
the lack of a cast. But then we have the case with an even number of
elements, which can either return 'the element near the middle' (i.e. an
ElementT) or the mean of the two middle-near elements (our solution now,
returns gul::statistics_result_type.

In most statistical applications the later is used, and so we do it here
now.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>