span: Add a backport of C++20's std::span as gul::span
[why]
std::span is the moral equivalent of std::string_view for
contiguous arrays of data. It has the potential of eliminating
function signatures of the type (pointer, size) in favor of more
type- and bounds-safe signatures.

[how]
Add an implementation of std::span by Tristan Brindle
(https://github.com/tcbrindle/span). Modify namespace to gul::
and eliminate unnecessary macros (e.g., we require C++14 for
GUL and need not test for it). Add only a minimal unit test:
We implicitly treat our backports as if they were standard
library code.