version: Return C strings instead of string_views
[why]
When we have potentially different versions of string_view, it can be
hard(er) to access the version information if one is not sure which
version is actually linked.
This is kind of hen and egg problem, you need to know the GUL version to
access the information about the GUL version...

[how]
Use C strings instead, which can always be accessed easily.

[note]
This is an API breaking change! BUT... When we made all symbols hidden
by default, we forgot to make the version string visible explicitely. So
the breaking change already happened some time ago. This adds no more
breaking to it :->

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>