meson: Install catch helper header in gul/
[why]
Prior to the renaming the catch2 test suite could be included by
  #include <gul/catch.h>

After we renamed gul to gul14 it has to be accessed by
  #include <gul14/catch.h>

To allow soft transistions for the projects that utilize gul, we
invented a new header file that (still) resides in gul/ and is
names catch.h, but just redirects to gul14/catch.h

But both catch.h files are currently written to gul14. This needs
special handling.

[how]
Add a special header install clause just for the compatibility catch.h.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>