deb: Add compatibility link libgul.so
[why]
We have a lot of projects that use -lgul in their Makefiles directly.
As GUL is renamed from gul to gul14 they would have to change their
linker flag.

Using pkg-config would prevent that, but it is not widely enough used.

So we add a symlink for their sake instead.

[note]
We do not use a debian/*.links file because that would have to be
configured with the package name etc pp...

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>