meson: Repair self containment tests
[why]
On darwin the build fails, because lib_ignore_me.a is empty.
It is also empty on other platforms, but noone cares ;)

[how]
With commit
  e3c41ccee meson: Use meson's project_name as include directory name
the content of the standalone_headers array changed from full-include-path
to just filesnames in 'gul14' subdirectory.

But the evaluation and creation of self-containment tests needs to adapt
to this and prepend the #include statement accordingly.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>