meson: Set Darwin compatibility version numbers
[why]
When building on Darwin it is impossible to build a GUL version that
has any patches on top of a tagged version. This happens only when
already 3 numbers are used (i.e. Doocs packet mode):
  libgul14.so.18.11.7.p5

The p5 can not be translated into a valid Darwin version number.

[how]
Just ignore the patch level in that cases. This uses a meson mechanism
introduced with meson 0.48. A warning will be issued on older
mesons, but builds are possible with the previous behavior/limitations.

[note]
Darwin also limits the numerical range of the version number parts, which
are not checked. So builds might still fail in that cases.The ranges are
wide enough for our typical use, though.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>