meson: Create compatibility link with correct suffix
[why]
The dynamic library file link libgul->libgul14 has to use the correct
platform dependent suffix (i.e. so, dylib, dll).

[how]
Unfortunately there is no mechanism forseen to query the suffix after
Meson found out which to use. So we just use the suffix rules as
detailed in Meson's documentation and do a very shallow check if
we could be right afterwards.

[note]
On Windows we do not create any softlink, as different versions of a
library all installed and resolved by softlinks is not supported.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>