deb: Resolve debian package update issues
[Why]
With commit #a234bcb (deb: Add old gul-dev packets to conflicts)
we ensured that we can not installed dev-doocs-libgul and
dev-doocs-libgul14 at the same time. But this has the side-effect
that we can't update the packages before removing the old package
first.

[How]
By adding the Replaces keyword, we help the package manager to
resolve the issue in case of an update.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>