Remove compat helper to use 'gul' instead of 'gul14'
[why]
We have a lot of problems with the compatibility links and files, that
were introduced to compile client code unchanged after the GUL renaming
from gul to gul14.
To make that possible some common-to-both-versions file are needed.
These files necessitate that the packets conflict each other and can not
be installed concurrently.
The conflicting leads to a lot of issues as the packet needs to be
removed/replaced at the exact right time in a CI build...

[how]
This commit dispenses with the compatibility idea altogether. libgul
and libgul14 are now two disjunct libraries. They can be used in
parallel, can be installed in parallel.
Client code has to be adapted. But that can happen slowly over time, and
after all (relevant) client code has been changed the libgul packets can
be removed from the development systems.

[note]
These are the more or less completely reverted commits:

05377d72   deb: Resolve debian package update issues
54da2ea8   meson: Create compatibility link with correct suffix
4e1a815a   meson: Add compatibility link libgul.so
a234bcbc   deb: Add old gul-dev packets to conflicts
3fa6ed83   meson: Install catch helper header in gul/
e8049284   Add compatibility header gul/catch.h
2502b69e   Add a test to ensure that GUL14 symbols can be accessed via namespace gul::
23599b0b   Documentation: Add remarks on alias namespace gul::

Signed-off-by: Fini <ulf.fini.jastrow@desy.de>