optional: Fix include guard
[why]
The (classic) include guard uses a quite common name.
Furthermore it uses the same name as gul-without-number includes.

This clashes if for some obscure reason gul-dev and gul14-dev are
installed concurrently on a system and both are somehow included.

[how]
Use unique name for include guard as recommended in Core Guideline SF.8.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>