meson: Do not append patchlevel to library files
[why]
The patchlevel does not belong to the library files (e.g. so or dll).
It is purely a packeting version number to replace previous library files.

This is a conceptual error.

[how]
Just do not add the patch version number to the library version number.

[note]
Also rewrote the documentation of the packet versioning description, so that
what happens why becomes (hopefully) more obvious.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>