meson: Simplify git_patch version generation
[why]
In some places git_version is set to empty string if something prevents us
from finding a meaningful patchlevel number.

Later, when git_version is used it can actually handle itself an unset
variable, so we do not need to handle it in a distributed way.

This makes the code more readable. Maybe. At least for me.

[how]
Just drop the needless code; checking for unset variable is already in.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>