meson: Reformat build description files
[why]
We use the free-indet method, where arguments on the next line are lined
up vertically to the first argument. This style has been used in early
examples of meson.build files. I though that ok, because this is the
style also used in a lot functional programming languages and thus not
weird to me.

But this is a C++ project, and here the typical indentation style is a
fixed number of blancs. I believe it is more readable for people used to
this style if we swap.

Also minor comment changes and replacement of 'if' with '?:'.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>