meson: Avoid double compilation
[why]
The sources are compiled twice, once for the shared and once for the
static library. Meson's reason for this is that shared libraries have to
be compiled position independent (-fpic), while static libraries can use
fixed positions. We never used that feature; although a user could have
enabled it with the Meson configuration switch:
  b_staticpic [true, false] Build static libraries as position independent
(Default is true)

Per default produce the same objects twice, both times with -fpic set.

[how]
Use the objects that we generate for the shared library also for the
static library, if possible.

[note]
I noticed this behavior when I changed the clientlib from using the
newer both_libraries() target to the 0.40.0 conform shared_library() and
static_library() targets.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>