statistics: Add minimum() and maximum()
[why]
Sometimes only the minimum or maximum value is needed,
not both. In these cases min_max() is too cumbersome
and might have a slight performance penalty.

Note that the names min() and max() cannot be used
because the 2-iterator version overloads would clash
with the existing gul14::min() and gul14::max().

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>