meson: Modernize pkg.generate() call
[why]
The way we generate the pkgconfig file has been obsoleted:

  DEPRECATION: Library gul14 was passed to the "libraries" keyword
  argument of a previous call to generate() method instead of first
  positional argument. Adding libgul14 to "Requires" field, but this
  is a deprecated behaviour that will change in a future version of
  Meson. Please report the issue if this warning cannot be avoided
  in your case.

[how]
The former way was to specify libraries with the libraries kwarg. But
now we have a positional argument that is smarter. Well. We do not use
that features. But now the old way became obsoleted...

pkg.generate(libraries : libgul, ...

becomes

pkg.generate(libgul, ...

It has some implications / improvements, but not here in this case in GUL.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>