meson: Fix debian package version number for clean releases
[why]
When building a clean release version (i.e. HEAD is tagged with the DOOCS
release tag, no additional commits after the tag) we get a strange zero
attached to the version string.

This does only show up in the debian/changelog, not in the diagnostic Meson
output:

  Message: Debian package names    doocs-libgul14-20-6-1, dev-doocs-libgul14
  Message: Shared library version  20.6.1

  doocs-libgul14-20-6-1 (20.6.10) stable; urgency=low
  * Bump version number to 2.3 and add DOOCS version information  (HEAD, tag: v2.3, tag: DOOCSVERSION_20_6_1)
 -- Tim Wilksen <tim.wilksen@desy.de>  Wed, 24 Jun 2020 12:58:49 +0200

20.6.1 vs 20.6.10 in the changelog.

The reason is that we attach the number of patchlevels EVEN IF IT IS ZERO.
This is not intended, usually a '0' should be replaced by ''.

[how]
Set the git_patch variable only if there is a patch applied on top of the
release tag.

Also add a diagnostic Meson output that shows the package version number
after configuring.

For example:
  Message: LIBGUL_API_VERSION       2.3
  Message: GIT FULL API VERSION    v2.3-2-gbe7fc4c-dirty
  Message: GIT VERSION TAG         DOOCSVERSION_20_6_1
  Message: Debian package names    doocs-libgul14-20-6-1, dev-doocs-libgul14
  Message: Debian package version  20.6.1.p2
  Message: Shared library version  20.6.1

Second line from bottom is new.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>