Revert "meson: Add rpath support" (partially)
This (almost) reverts commit a179dc519cf3f30d16d296d68bab19aae2c4f9d1.

[why]
The RPATH is set in the library object. This does not make much sense.

The RPATH is used to find libraries at runtime. GUL does not depend on
any (nonstandard) library, especially not a library that is build by GUL
itself. It would be different when GUL has a sub-library GUL-helper.so
or something. That (sub) library would benefit from the added RPATH.

But assume an executable that uses (links) GUL. The loader needs to find
all needed libraries, and builds up the RPATH by all RPATH elements in
the libraries it links and of course most important the RPATH entry of
the actual executable. But: When the GUL.so has been found it ... well
has been found already, obviously. Only after GUL has been found the
RPATH the (reverted) commit adds to the library object can be evaluated
and followed. But ... we already have it, or we do not have it, but then
the library's RPATH can not be know, right?

RPATH does make sense to find libraries that WE depend on. Not ourselves.

What does make sense is to add the path to GUL when building an
executable (or library) that depends on GUL. This is handled by the
pkgconfig entry - thus it is kept from the (reverted) commit in question.

[how]
Do not add RPATH entries to the library object.
Do add (keep) path entries to the pkgconfig specification.

[note]
I guess the (reverted) commit is inspired by / a copy of similar commits
done to the DOOCS libraries. That libraries depend on sub-libraries, and
there an RPATH entry in the library object does make sense. How it is
set there is questionable (i.e. wrong), but works in usual usecases.
But here it is not appropriate imho.